Add an example for a CLI app also used as a Hera script (#1536)
**Pull Request Checklist**
- [x] Related to #1530
- [x] Tests added
- [x] Documentation/examples added
- [x] [Good commit messages](https://cbea.ms/git-commit/) and/or PR
title

**Description of PR**
Currently, a user may want to reuse a CLI app with Hera. `click` is a
popular choice but garbles the underlying function, rendering it
unusable with the `script` decorator (and not worth first-party support
at this time). Cappa, used for the `hera` CLI itself, does not interfere
with the function, so it is usable with the script decorator.

This PR adds a use-case/integration example for reusing a CLI app with
Hera. Also pointing out in a warning box that `click` will not
ordinarily work with the `script` decorator (requiring the `Script`
class).

cc @DanCardin sharing the cappa love 🫶

Signed-off-by: Elliot Gunton <elliotgunton@gmail.com>
Co-authored-by: Sambhav Kothari <skothari44@bloomberg.net>